rename: weatherBoardRain -> weatherBoardPrecipiation 파일명 변경
