+++ frontend/src/pages/weatherBoardPage/components/weatherBoardPrecipiation/WeatherBoardPrecipiation.tsx
